Continued testing

+++ Checkout/Entities/Product.cs
@@ -2,12 +2,12 @@ namespace Checkout.Entities;
 
 public class Product
 {
-    public Product(string sku, decimal price )
+    public Product(string sku, double price )
     {
         this.Sku = sku;
         this.Price = price;
     }
 
     public string Sku { get; set; }
-    public decimal Price { get; set; }
+    public double Price { get; set; }
 }
